Changed int to double

+++ Labb3_CalorieTrackerMongoDB/Models/WeeklySummary.cs
@@ -8,14 +8,14 @@ namespace Labb3_CalorieTrackerMongoDB.Models
         public bool HasLog { get; set; }
 
         public int GoalCalories { get; set; }
-        public int GoalProtein { get; set; }
-        public int GoalCarbs { get; set; }
-        public int GoalFat { get; set; }
+        public double GoalProtein { get; set; }
+        public double GoalCarbs { get; set; }
+        public double GoalFat { get; set; }
 
         public int ActualCalories { get; set; }
-        public int ActualProtein { get; set; }
-        public int ActualCarbs { get; set; }
-        public int ActualFat { get; set; }
+        public double ActualProtein { get; set; }
+        public double ActualCarbs { get; set; }
+        public double ActualFat { get; set; }
 
         public int CaloriesDiff => HasLog ? ActualCalories - GoalCalories : 0;
